@@ -4,7 +4,7 @@ version: 0.0.1
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.16.0"
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
@@ -15,39 +15,4 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^3.0.0
 
-
-
-
-
 flutter:
-
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
-  
